+++ backend-test/RequirementTests/PreProcessingTests.hs
@@ -7,14 +7,15 @@ module RequirementTests.PreProcessingTests
 ( preProcTestSuite ) where
 
 import Data.Text as T hiding (map)
-import Test.HUnit (Test (..), assertEqual)
+import Test.Tasty (TestTree, testGroup)
+import Test.Tasty.HUnit (assertEqual, testCase)
 import Text.HTML.TagSoup (Tag (..))
 import WebParsing.PostParser (pruneHtml)
 
-createTest :: (Eq a, Show a, Eq b, Show b) => (a -> b) -> String -> [(a, b)] -> Test
-createTest function label input = TestLabel label $ TestList $ map (\(x, y) ->
-                                TestCase $ assertEqual ("for (" ++ show y ++ ")")
-                                y (function x)) input
+createTest :: (Eq a, Show a, Eq b, Show b) => (a -> b) -> String -> [(a, b)] -> TestTree
+createTest function label input = testGroup label $ Prelude.zipWith (\(x :: Int) (y, z) ->
+                                testCase ("Test " ++ show x) $ assertEqual ("for (" ++ show z ++ ")")
+                                z (function y)) [0..] input
 
 
 pruneHtmlInputs :: [([Tag T.Text], [Tag T.Text])]
@@ -33,9 +34,9 @@ pruneHtmlInputs = [
     )
     ]
 
-pruneHtmlTests :: Test
+pruneHtmlTests :: TestTree
 pruneHtmlTests = createTest pruneHtml "filtering out html attributes" pruneHtmlInputs
 
 -- functions for running tests in REPL
-preProcTestSuite :: Test
-preProcTestSuite = TestLabel "Pre-processing tests" $ TestList [pruneHtmlTests]
+preProcTestSuite :: TestTree
+preProcTestSuite = testGroup "Pre-processing tests" [pruneHtmlTests]