+++ backend-test/RequirementTests/RequirementTests.hs
@@ -8,12 +8,12 @@ Module that contains the test suites for all the requirement tests.
 module RequirementTests.RequirementTests
 (  requirementTests  ) where
 
-import Test.HUnit (Test (..))
 import RequirementTests.ModifierTests (modifierTestSuite)
 import RequirementTests.PostParserTests (postTestSuite)
 import RequirementTests.PreProcessingTests (preProcTestSuite)
 import RequirementTests.ReqParserTests (reqTestSuite)
+import Test.Tasty (TestTree, testGroup)
 
 -- Single test encompassing all requirement test suites
-requirementTests :: Test
-requirementTests = TestList [reqTestSuite, postTestSuite, preProcTestSuite, modifierTestSuite]
+requirementTests :: TestTree
+requirementTests = testGroup "Requirements" [modifierTestSuite, preProcTestSuite, postTestSuite, reqTestSuite]